requirements.txt
Flask==2.3.2
pdf2image==1.16.3
Pillow==10.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image
pdfminer.six
gunicorn